pyproject.toml
[project]
name = "django-postgres-rls"
version = "0.1.0"
description = "Django middleware for PostgreSQL Row-Level Security using role switching"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["django", "postgresql", "rls", "row-level-security", "middleware"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "django>=3.2,<6.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.0",
    "pytest-cov>=4.0",
    "tox>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "testcontainers[postgres]>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/django-postgres-rls"
Repository = "https://github.com/yourusername/django-postgres-rls"
Documentation = "https://github.com/yourusername/django-postgres-rls#readme"
"Bug Tracker" = "https://github.com/yourusername/django-postgres-rls/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["django_postgres_rls"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "tox>=4.32.0",
    "testcontainers[postgres]>=4.0.0",
]
